requirements.txt
Bio==1.3.3
CairoSVG==2.5.2
lxml==4.6.3
matplotlib==3.3.4
matplotlib_venn==0.11.6
numpy==1.20.1
pandas==1.2.4
Pillow==9.0.0
rdkit==2009.Q1-1
rdkit_pypi==2021.9.4
requests==2.25.1
scipy==1.6.2
seaborn==0.11.1
statannot==0.2.3
statsmodels==0.12.2
streamlit==1.4.0
tqdm==4.59.0
